remove default server provided to connections
